Main: Moved Arena to views / added useWebSocket hook / using babel-eslint
