+++ package.json
@@ -25,38 +25,39 @@
     "clean": "rm -rf dist/"
   },
   "devDependencies": {
-    "@babel/cli": "^7.8.3",
-    "@babel/core": "^7.9.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.9.5",
+    "@babel/cli": "^7.10.1",
+    "@babel/core": "^7.10.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.10.1",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.9.5",
-    "@babel/preset-react": "^7.9.4",
-    "@babel/register": "^7.9.0",
-    "@babel/runtime-corejs3": "^7.9.2",
+    "@babel/preset-env": "^7.10.2",
+    "@babel/preset-react": "^7.10.1",
+    "@babel/register": "^7.10.1",
+    "@babel/runtime-corejs3": "^7.10.2",
+    "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "chai": "^4.2.0",
     "cross-env": "^7.0.2",
     "css-loader": "^3.5.2",
-    "eslint": "^7.0.0",
+    "eslint": "^7.2.0",
     "eslint-config-airbnb": "^18.1.0",
     "eslint-config-prettier": "^6.10.1",
     "eslint-import-resolver-webpack": "^0.12.1",
     "eslint-plugin-chai-friendly": "^0.6.0",
     "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jsdoc": "^25.4.2",
+    "eslint-plugin-jsdoc": "^27.0.4",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-mocha": "^7.0.0",
+    "eslint-plugin-mocha": "^7.0.1",
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-react": "^7.20.0",
-    "eslint-plugin-react-hooks": "^4.0.2",
+    "eslint-plugin-react-hooks": "^4.0.4",
     "git-revision-webpack-plugin": "^3.0.4",
     "google-fonts-plugin": "^5.0.2",
     "html-replace-webpack-plugin": "^2.5.6",
     "html-webpack-plugin": "^4.2.0",
     "husky": "^4.2.5",
-    "lint-staged": "^10.2.4",
+    "lint-staged": "^10.2.9",
     "mini-css-extract-plugin": "^0.9.0",
-    "mocha": "^7.1.1",
+    "mocha": "^7.2.0",
     "mochapack": "^2.0.0",
     "node-dev": "^4.0.0",
     "node-sass": "^4.13.1",
@@ -101,14 +102,14 @@
     "express-fileupload": "^1.1.7-alpha.3",
     "express-session": "^1.17.0",
     "file-saver": "^2.0.2",
-    "i18next": "^19.4.1",
+    "i18next": "^19.4.5",
     "is-valid-coordinates": "^1.0.0",
     "jsonwebtoken": "^8.5.1",
     "jwk-to-pem": "^2.0.3",
-    "log4js": "^6.1.2",
+    "log4js": "^6.3.0",
     "marked": "^1.1.0",
     "node-schedule": "^1.3.2",
-    "node-stream-zip": "^1.11.1",
+    "node-stream-zip": "^1.11.2",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
     "pg": "^8.2.1",
