+++ webapp/loggedin/appView.js
@@ -1,6 +1,7 @@
 import './appView.scss'
 
 import React from 'react'
+import { useLocation } from 'react-router'
 
 import AppHeader from './appHeader/appHeader'
 import AppSideBar from './appSideBar/appSideBar'
@@ -8,9 +9,9 @@ import AppJobMonitor from './appJob/appJobMonitor'
 import AppErrors from '../app/appErrors/appErrors'
 import LoggedInModuleSwitch from './modules/loggedInModuleSwitch'
 
-const AppView = (props) => {
-  const { location } = props
-  const pathName = location.pathname
+const AppView = () => {
+  const location = useLocation()
+  const { pathName } = location
 
   return (
     <>
