+++ webapp/views/Arena/index.js
@@ -1,0 +1,1 @@
+export { default } from './Arena'
